[threading] Fix a race condition between managed suspend/abort and GC STW.
STW and managed suspend share a lock to ensure there is only one active suspend initiator.

This originally caused a problem in the managed suspend path as it could be async suspended while it was
in the middle of the suspend protocol (check if top of stack is managed, install abort handlers, etc).

This was worked around by putting that inside a suspend critical section. The resulting code was horrible
as suspend boundaries became unstructured (yeah, pretty stupid idea on my end). Not only that but it had
a race condition.

If a thread doing managed suspend raced with a STW and lost, it would block on the global suspend lock to
finish. That thread would be blocked inside a critical section - the single thing you must not do inside one.

This could be reproduced by stressing mono/tests/sgen-new-threads-dont-join-stw.

The solution is to execute the managed suspend sequence in the same fashion as of a STW. IOW, take locks,
suspend, do stuff, maybe resume and drop locks.

The critical part is the "do stuff" one, which is different for abort and suspend, but everything else
is the same. So we wrap that into a callback and put the common parts in mono-threads. This gives
us structured locking and suspend boundaries.

This won't cause a GC to take longer to init as the critical section would block the  STW from progressing
anyways.


Commit migrated from https://github.com/mono/mono/commit/3af402a0d048b568216689dc5cac755d275ce5d1