[aot] Added JSON dumping support.
This is enabled by passing "aot=dump" to the Mono AOT driver and will create a JSON file in the working directory with the name of the assembly.

It's still a work-in-progress and will see improvements in the data dumped as it's needed.


Commit migrated from https://github.com/mono/mono/commit/e3279063de26133a963472eb0b8b3bb60d403d3b