[Socket] Add DualMode property and adjust to Mono semantics (continuation to mono/mono#20048)
The new socket constructor that only takes a SocketType and a
ProtocolType defaults to InterNetworkV6 (this was done on a previous patch).

In addition, we explicitly set "DualMode" to true, so that on Windows,
this gets set explicitly (it does not matter on Unix).

We surface the DualMode property to allow servers that want to just
bind to the IPv6 address to do so by creating the socket and then
changing DualMode to false.

This is *not* the same setting that Windows has, but it is the
semantics that Mono applications depend on.  So a test has been added
to ensure that we do not regress.

This also switches to use SocketOptionName from the ReferenceSources.


Commit migrated from https://github.com/mono/mono/commit/53dc56ee39a8e3b013231957aca4671b202c6410