[threadpool-ms-io] Remove socket synchronously from backend when closing socket
To achieve this, we need to wait for the update to be applied to the backend.

This is mainly for the epoll backend, as we do not need to remove the fd from the backend when using kqueue.


Commit migrated from https://github.com/mono/mono/commit/fa2649301755019b14cea27d764ba9a7b3376632