[utils] Change mono_conc_hashtable write operations to rely on external locking.
The original design was to pass a lock to the mono_conc_hashtable constructor and
then the write operations would do locking themselves.

That turned out to be a problem for two reasons:

First, it hidden the fact that locking was happening. Every other data structure in
the runtime rely on the caller to perform locking - thus that design ended up been
a lie.

Second, under coop, taking the lock became a problem as leak locks should be taken
in GC Safe mode and other locks should be taken under GC Unsafe mode. Taking a lock
in the wrong GC mode makes the suspend watchdog bark really loudly.

The runtime, funny enough has 2 usages of the conc hashtable, one with the image lock
and one with the domain lock - and each must be taken in a different GC mode. That
meant that either overcomplicating the design, or just removing locking from it.

This commit does that. It removes the mutex from mono_conc_hashtable and hoist locking
around the calls that used to do it.


Commit migrated from https://github.com/mono/mono/commit/d164e49e079a95ae8998f2e487a8a841fd9c0148