[verifier] Fix the verifiers to handle Roslyn style fixed blocks.
Roslyn used byref vars whenever possible to store fixed statements. This leads to confusion in
both the full verifier and the JIT built-in one.

Now we skip it in the JIT one and allow it when not in strict mode for the full verifier.


Commit migrated from https://github.com/mono/mono/commit/72a7e8de9798526c97a74dba2d9866aa9182cf84