[sgen] Fix mod union cardtable marking race
When the nursery collector is starting and we have a concurrent collection in progress we need to save the global card table entries to the mod union card tables. When doing so we were marking all the mod union cards with either 0 or 1, so we could lose a mark if the concurrent collector tries to mark the same card at the same time, since the concurrent collector now runs concurrent even with nursery collections. We fix this by never marking with 0.


Commit migrated from https://github.com/mono/mono/commit/e3e44c0a1351ff2a519411bc567b28b0da66f7a8