[runtime] Move pending_exception from InternalThread to Thread to avoid crashing on domain unload.
InternalThread is a fundamentally different type WRT to appdomain. All instances of Thread bound to the same
os thread have a reference to a shared InternalThread.

InternalThread is always allocated in the root domain, so it MUST NOT have references to other domains as
it will lead to crashes on unload.

This was the case with pending_exception. It could point to an exception meant to the domain unloading.

This bug became more and more frequent as we move to use mono_set_pending_exception in more places. I could
reproduce it in 10 minutes of looping appdomain-unload.exe.


Commit migrated from https://github.com/mono/mono/commit/15f5673e36f6e9817906a9dc135fdda2b2c41a78